Reconfigure delete survey and related questions success

+++ src/main/java/net/yorksolutions/douarmouacapstone1be/Services/SurveyService.java
@@ -2,7 +2,9 @@ package net.yorksolutions.douarmouacapstone1be.Services;
 
 import net.yorksolutions.douarmouacapstone1be.DTOs.NewSurveyRequestDTO;
 import net.yorksolutions.douarmouacapstone1be.DTOs.UpdateSurveyRequestDTO;
+import net.yorksolutions.douarmouacapstone1be.Entities.Question;
 import net.yorksolutions.douarmouacapstone1be.Entities.Survey;
+import net.yorksolutions.douarmouacapstone1be.Repositories.QuestionRepository;
 import net.yorksolutions.douarmouacapstone1be.Repositories.SurveyRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -14,9 +16,11 @@ import java.util.Optional;
 public class SurveyService {
 
     SurveyRepository surveyRepository;
+    QuestionRepository questionRepository;
 
-    public SurveyService(SurveyRepository surveyRepository) {
+    public SurveyService(SurveyRepository surveyRepository, QuestionRepository questionRepository) {
         this.surveyRepository = surveyRepository;
+        this.questionRepository = questionRepository;
     }
 
     public Survey createSurvey(NewSurveyRequestDTO requestDTO){
@@ -43,6 +47,13 @@ public class SurveyService {
     }
 
     public void deleteSurvey(Long surveyId){
+        Optional<Survey> surveyOptional = this.surveyRepository.findById(surveyId);
+        if (surveyOptional.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
+        Iterable<Question> surveyQuestions = this.questionRepository.findAllBySurveyOwner(surveyOptional.get());
+        this.questionRepository.deleteAll(surveyQuestions);
         this.surveyRepository.deleteById(surveyId);
     }
